events: make models no_std-ready and align timestamps
- make `radroots-events` no_std/alloc-friendly with gated serde/typeshare/ts-rs derives and core feature propagation
- add a typeshare-only kinds view and refresh generated TS bindings
- compat: post/profile published_at now u32; codecs and nostr adapters use saturating timestamp conversion

+++ events/src/post.rs
@@ -1,11 +1,14 @@
 use crate::RadrootsNostrEvent;
-use serde::{Deserialize, Serialize};
 #[cfg(feature = "ts-rs")]
 use ts_rs::TS;
 
+#[cfg(not(feature = "std"))]
+use alloc::string::String;
+
 #[cfg_attr(feature = "ts-rs", derive(TS))]
 #[cfg_attr(feature = "ts-rs", ts(export, export_to = "types.ts"))]
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[derive(Clone, Debug)]
 pub struct RadrootsPostEventIndex {
     pub event: RadrootsNostrEvent,
     pub metadata: RadrootsPostEventMetadata,
@@ -13,18 +16,20 @@ pub struct RadrootsPostEventIndex {
 
 #[cfg_attr(feature = "ts-rs", derive(TS))]
 #[cfg_attr(feature = "ts-rs", ts(export, export_to = "types.ts"))]
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[derive(Clone, Debug)]
 pub struct RadrootsPostEventMetadata {
     pub id: String,
     pub author: String,
-    pub published_at: u64,
+    pub published_at: u32,
     pub kind: u32,
     pub post: RadrootsPost,
 }
 
 #[cfg_attr(feature = "ts-rs", derive(TS))]
 #[cfg_attr(feature = "ts-rs", ts(export, export_to = "types.ts"))]
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[derive(Clone, Debug)]
 pub struct RadrootsPost {
     pub content: String,
 }
