+++ events/src/profile.rs
@@ -1,11 +1,14 @@
 use crate::RadrootsNostrEvent;
-use serde::{Deserialize, Serialize};
 #[cfg(feature = "ts-rs")]
 use ts_rs::TS;
 
+#[cfg(not(feature = "std"))]
+use alloc::string::String;
+
 #[cfg_attr(feature = "ts-rs", derive(TS))]
 #[cfg_attr(feature = "ts-rs", ts(export, export_to = "types.ts"))]
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[derive(Clone, Debug)]
 pub struct RadrootsProfileEventIndex {
     pub event: RadrootsNostrEvent,
     pub metadata: RadrootsProfileEventMetadata,
@@ -13,18 +16,20 @@ pub struct RadrootsProfileEventIndex {
 
 #[cfg_attr(feature = "ts-rs", derive(TS))]
 #[cfg_attr(feature = "ts-rs", ts(export, export_to = "types.ts"))]
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[derive(Clone, Debug)]
 pub struct RadrootsProfileEventMetadata {
     pub id: String,
     pub author: String,
-    pub published_at: u64,
+    pub published_at: u32,
     pub kind: u32,
     pub profile: RadrootsProfile,
 }
 
 #[cfg_attr(feature = "ts-rs", derive(TS))]
 #[cfg_attr(feature = "ts-rs", ts(export, export_to = "types.ts"))]
-#[derive(Clone, Debug, Serialize, Deserialize)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[derive(Clone, Debug)]
 pub struct RadrootsProfile {
     pub name: String,
     #[cfg_attr(feature = "ts-rs", ts(optional, type = "string | null"))]
